Item drop progress
Made ItemDrop animated. Added logic to physics to manage item pickup. Added some code to World so that we spawn a few test items on startup.

Note: none of this is tied into the inventory system yet, we just have the process of moving an item toward the player then removing it from the scene. We'll need to add logic to add items to the player's inventory when they're picked up.