Add devmode
Added developer mode to options menu and game settings to be able to turn features on/off without changing code. Currently it is just used to disable light timer/game over code. We can expand this to change other behavior as well.

TODO: Remove before we ship the game. Don't want to leave in any hax ;)