Add starter journal code and fix menu display logic
Edited game menus logic so we can better control escaping out of menus, support pausing on the inventory and journal menus, etc.

Added journal item to the game, which can be picked up and put into inventory. After which, the journal can be accessed through the inventory at any time- we could probably bind it to "J" as well for ease of access.

Still working on actually putting text in the journal- planning to have journal pages that we can create and inject into the journal whose load is controlled by the page prev/next buttons.