Handle journal page picked up before journal
If a journal page is found before the journal, it will appear in the inventory as an unknown journal page. Once the journal is picked up, the page will be unlocked to read in the journal (the logic here being that the player needs the context of the journal to use the recipe).