Add gate logic
Added gate and gate interact logic. If the player interacts using 'spacebar' near the gate, they will get a text dialogue. If it's locked, they'll get the locked dialogue. They can unlock the gate if they have the item id that is specified as the key for that gate. When the user interacts with that item in their inventory, they unlock the gate and a new dialogue is shown.

We can update this later to have the user select a specific item from the inventory rather than just "if it's in the inventory".